Expand quick crimes when editing
When a user edits the quick crimes, if the bar is collapsed, it may not be obvious that the crimes are being added.

Solution initially was to refactor the container generation in `content` to add a toggle function so it would be easy to generically toggle any collapsible container box. Instead, a simpler solution was to collapse the box directly. A bit leaky, but this behaviour is so far specific to the crimes box.