Add racing filter for custom races (#892)
* Add racing filter

* Make the racing filter compact

* fix pull request remarks

* Create multi-select dropdown menu

I created a multi-select dropdown menu and implemented it in the racing filter.
Only tested on desktop, but should also work on mobile devices.

* Change multi-select dropdown to checkbox scroll area

I couldn't fix the jump to top bug on the multi-select dropdown, so i changed it to a checkbox scroll area.
This will save space by not having a checkbox for every track visible at once on the filter.
Again, hasn't been checked on mobile device.